# swagger.yaml
openapi: 3.0.3
info:
  title: InternFlare Job API
  version: 1.0.0
  description: API documentation for internflarejobs endpoints protected with JWT authentication.
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Job:
      type: object
      required:
        - jobTitle
        - jobDesc
        - jobType
        - stillOpened
        - companyId
      properties:
        jobTitle:
          type: string
          example: "Frontend Developer"
          description: "Title of the job"
        jobDesc:
          type: string
          example: "Develop and maintain web applications"
          description: "Detailed description of the job"
        jobType:
          type: string
          example: "Full-time"
          description: "Type of job (e.g., Full-time, Part-time, Contract)"
        stillOpened:
          type: boolean
          example: true
          description: "Whether the job is still open"
        companyId:
          type: string
          example: "60d5ecb74b24c72b8c8b4567"
          description: "ID of the company posting the job"
...
  responses:
    UnauthorizedError:
      description: "Authentication information is missing or invalid"
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "fail"
              message:
                type: string
                example: "You are not logged in!"
paths:
  /api/v1/jobs:
    get:
      security:
        - bearerAuth: []
      summary: "Get all jobs"
      description: "Retrieves a list of all jobs with optional filtering, sorting, and pagination."
      parameters:
        - in: query
          name: sort
          schema:
            type: string
          description: "Sort order for the jobs"
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: "Page number for pagination"
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: "Number of jobs per page"
      responses:
        "200":
          description: "A list of jobs"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  results:
                    type: integer
                    example: 2
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      security:
        - bearerAuth: []
      summary: "Create a new job"
      description: "Creates a new job with the provided details."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Job"
      responses:
        "201":
          description: "Job created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      newJob:
                        $ref: "#/components/schemas/Job"
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "fail"
                  message:
                    type: string
                    example: "no job input"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...